+++ domain/src/main/kotlin/gr/tsambala/tutorbilling/domain/di/DomainModule.kt
@@ -6,6 +6,7 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import gr.tsambala.tutorbilling.data.dao.LessonDao
 import gr.tsambala.tutorbilling.data.repository.StudentRepository
+import gr.tsambala.tutorbilling.data.repository.TutorBillingRepository
 import gr.tsambala.tutorbilling.domain.lesson.*
 import gr.tsambala.tutorbilling.domain.student.*
 import javax.inject.Singleton
@@ -30,15 +31,18 @@ object DomainModule {
 
     @Provides
     @Singleton
-    fun provideLessonUseCases(dao: LessonDao): LessonUseCases =
+    fun provideLessonUseCases(
+        dao: LessonDao,
+        repository: TutorBillingRepository
+    ): LessonUseCases =
         LessonUseCases(
             getAllLessons = GetAllLessons(dao),
             getLessonById = GetLessonById(dao),
-            getLessonsByStudentId = GetLessonsByStudentId(dao),
+            getStudentLessons = GetStudentLessons(repository),
             getLessonsWithStudents = GetLessonsWithStudents(dao),
             getLessonsWithStudentsByStudentAndDateRange = GetLessonsWithStudentsByStudentAndDateRange(dao),
-            insertLesson = InsertLesson(dao),
-            updateLesson = UpdateLesson(dao),
+            addLesson = AddLesson(repository),
+            updateLesson = UpdateLesson(repository),
             deleteLesson = DeleteLesson(dao),
             updateLessonPaidStatus = UpdateLessonPaidStatus(dao)
         )