+++ domain/src/main/kotlin/gr/tsambala/tutorbilling/domain/lesson/AddLesson.kt
@@ -4,8 +4,8 @@ import gr.tsambala.tutorbilling.data.model.Lesson
 import gr.tsambala.tutorbilling.data.repository.TutorBillingRepository
 import javax.inject.Inject
 
-class UpdateLesson @Inject constructor(
+class AddLesson @Inject constructor(
     private val repository: TutorBillingRepository
 ) {
-    suspend operator fun invoke(lesson: Lesson) = repository.updateLesson(lesson)
+    suspend operator fun invoke(lesson: Lesson): Long = repository.addLesson(lesson)
 }